requirements.txt
mediapipe
opencv-python
numpy
livekit
python-dotenv
